kseed/pyproject.toml
[tool.poetry]
name = "kseed"
version = "0.1.0"
description = "Pulumi infrastructure as code for homelab k3s cluster"
authors = ["Homelab Owner <admin@homelab.local>"]
readme = "README.md"
packages = [{include = "kseed"}]

[tool.poetry.dependencies]
python = "^3.14"
python-semantic-release = "^9.8.0"
pulumi = "^3.140.0"
pulumi-kubernetes = "^4.18.0"
python-dotenv = "^1.0.0"
typer = {extras = ["all"], version = "^0.15.0"}
rich = "^13.9.0"
textual = ">=0.1.0"
pyyaml = "^6.0.2"
pydantic = "^2.10.0"
kubernetes = "^32.0.0"

[tool.poetry.scripts]
kseed = "kseed.cli.commands:app"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.semantic_release]
version_variable = [
    "kseed/__init__.py:__version__",
    "pyproject.toml:tool.poetry.version"
]
branch = "main"
build_command = "poetry build"
upload = false
changelog_file = "CHANGELOG.md"
hvcs = "github"
tag_commit = true
